+++ Routing/TagRouter.php
@@ -93,6 +93,10 @@ class TagRouter implements ChainedRouterInterface, RequestMatcherInterface
      */
     public function matchRequest(Request $request)
     {
+        if ($this->configResolver->getParameter('routing.enable_tag_router', 'eztags') === false) {
+            throw new ResourceNotFoundException('Config says to bypass TagRouter');
+        }
+
         $requestedPath = rawurldecode($request->attributes->get('semanticPathinfo', $request->getPathInfo()));
         $pathPrefix = $this->generator->getPathPrefix();